🚚 lib配下に移動させ、class化した
